managers to separate file
