+++ hire_hire/interview/models.py
@@ -1,21 +1,9 @@
-from random import sample
-
 from django.contrib.auth import get_user_model
 from django.db import models
 
+from interview.managers import QuestionManager
+
 User = get_user_model()
-
-
-class QuestionManager(models.Manager):
-    """
-    Кастомный менеджер для добавления метода
-    выборки данного кол-ва случайных вопросов.
-    """
-
-    def random(self, cnt):
-        ids = list(self.get_queryset().values_list('id', flat=True))
-        rand_ids = sample(ids, min(cnt, len(ids)))
-        return self.get_queryset().filter(id__in=rand_ids)
 
 
 class Language(models.Model):
